Revised AMI_Stergiou and RQA
Made first minimum and 1/5th values separate outputs in AMI_Stergiou. This makes using them easier in the BAR App. Commented out the plotting feature of the RQA code. This will be removed once it is fully operational in the BAR app. Added the weighted recurrence matrix to the output as well as the data used in the plot. The later helps make the plot in BAR as it may have been shortened, normalized, etc.